Bump reccmp to the fork's current master. Its index projection has no
schema field, so the loader rejects only documents that still carry one
(collected by an older reccmp). decomplint skips targets absent from the
index: the srEXT JPEG/Unzip targets are configured only when their pinned
vendor sources are mounted.

+++ src/wiz8/local_code/Magic.cpp
@@ -1640,7 +1640,7 @@ unsigned int GetBestSpellbookSkillForSpell(W8Character* character, int spell_id,
                 failure = 0;
             }
 
-            shortfall = static_cast<int>(GetMinimumCasterLevelForSpell(spell_id)) -
+            shortfall = GetMinimumCasterLevelForSpell(spell_id) -
                         GetTotalCasterLevel(character, book, 1) - 1 + power_level;
             if (shortfall > 0) {
                 failure += g_spell_records[spell_id].spell_level * shortfall;
